Add the repo sync step
Prior to this commit the readme did not contain instructions
for syncing a users fork and local copy to the upstream
project. This commit adds those instructions and includes
example output, except for retrieving/applying/pushing
changes because I do not have a fork to get those messages
from currently.