Add cleaning instructions
Each kata ends after the working branch has been merged and cleaned.
Prior to this commit there were no instructions for cleaning the
merged working branch from the users local and remote fork of the
project. Left in this state the user attempting to perform the kata
for a second time would run into errors.

This commit ensures that the cleaning instructions and examples are
included.